Ignore long lines with URLs
If a commit message has long lines that are long because of URLs that
cannot be split over multiple lines, ignore the rule for that line.

The URL matching regular expression is a basic implementation, but it
also doesn't need to be 100% perfect for this use case.

Use the regex and lazy_static crates that are already dependencies of
other dependencies. This shouldn't further increase the build size.