Decrease release sizes before publishing
Strip the symbols from the release builds before publishing. This saves
space in the final version. Some release builds shrink from 4MB to
700-900KB.

Users will have to download even less (compressed around 400KB max).

To strip the symbols from Linux builds (on a macOS host) I created
Docker images to help with this, and run the strip command in
containers. It's not possible to strip Linux targets on macOS.