Support Git commit-msg hook
Add support for the Git commit-msg hook format. When the `--file <FILE>`
option is given, it will read the file from the given path.

This commit format does not include any Git commit SHA information as
the commit itself hasn't been made yet at this time. For this reason I
made the sha fields optional on the Commit struct.

I also updated the error handling of the program now that the lint
command function has more complex scenarios to handle.