Support other Git cleanup modes
When the Git commit-msg hook is used the file from which we can read the
commit message has all the contents, including potential comments. To
ensure we don't validate these comments or other lines that should be
ignored, implement Git's clean up behavior.

Git has the option to configure how to clean up the commit message. From
the docs:

```
--cleanup=<mode>
    This option determines how the supplied commit message should be
    cleaned up before committing. The <mode> can be strip, whitespace,
    verbatim, scissors or default.

    strip
        Strip leading and trailing empty lines, trailing whitespace,
        commentary and collapse consecutive empty lines.
    whitespace
        Same as strip except #commentary is not removed.
    verbatim
        Do not change the message at all.
    scissors
        Same as whitespace except that everything from (and including)
        the line found below is truncated, if the message is to be
        edited. "#" can be customized with core.commentChar.
        # ------------------------ >8 ------------------------
    default
        Same as strip if the message is to be edited. Otherwise
        whitespace.
```

Source: https://git-scm.com/docs/git-commit

To know the config method and the `commit.commentChar` option, fetch
those config values before parsing the commit so they can be used to
clean up the commit beforehand.

Update the delimiter used to differentiate between commit messages,
because I used the scissors Git uses as well and added it to this
comment makes it think there are two commits.