Update terminology rule and violation
I first used RuleType and Validation, but there's no need to specify
type as it's already an enum of types.

Validation to violation. Validation implies that it's either good or
bad, and it's always bad so it's a violation.

Rename "kind" to "rule" in the Violation struct as it's which Rule the
violation was about.