Move rule and violation to their own file
Extract some things out to the rule file so that the commit file is a
bit more manageable. There's still too much lines of code there, but we
can start by extracting this.

This way all the match-statements for all the rule types are in one
place and easier to find. Especially for the String to Rule enum
matching, that one won't prompt a compiler warning if one rule is
missing.