Add whitespace rule, check for leading whitespace
I noticed that when I accidentally added a whitespace to the beginning
of subjects it would not fail on another rule like
SubjectCapitalization.

This was because the subject was trimmed at the start of the String and
hid any whitespace that was present. Removing this trim caused a couple
test failure to pop up where commits had leading whitespace, because of
a `git commit` command options issue (`-m message`, instead of
`-m message` option) and some escaping that was missing from multi line
Strings.