Fix Git's core.commentChar config reading
The implementation of the Git config core.commentChar was not correct.
It was a copy of the cleanup mode config fetch function.

Update the implementation and add tests. I tested it by checking which
rules were triggered and if comments after the scissors were seen as
part of the message or not. I couldn't implement tests in such a way for
just the `comment_char` function without breaking parallel testing.