Add support for fenced code block with language
When a code block has a language identifier, accept it as valid. This
should complete support for long lines in Markdown code blocks. Use
a regular expression to catch all the different whitespace scenarios
that can occur.

This change also accounts for leading whitespace for lines with a fenced
code block start and end.