Add linting by commit SHA
Allow users to specify a single commit SHA to lint. This means it
supports not just commit ranges, but single commits as well.

I wanted to add this to make it easier to lint specific commits without
always having to add a specific range. Previously without a range it
would lint from that commit to the rest of the history, like `git log`'s
behavior.