Merge subcommands into one main command
Previously I split the lint and lint hook commands into two subcommands,
but instead I merged them back. There is not a lot of functionality in
the commands themselves, so renaming the file option to
`--hook-message-file`, making it more explicit should help with any
confusion.

I also removed the `lint` command so that it doesn't even need a
subcommand. Now it's just a matter of running `gitlint` instead of
`gitlint lint`, which repeats itself a bit. I don't think there will be
any functionality other than linting. I will add branch name linting,
but that can be part of the commit linting process as well, and
potentially disabled with a flag.

Add verbatim_doc_comment because I realized the example formatting
wasn't how I intended it to look.