Update MessageLineLength rule to check all lines
Validate all lines and include them in the output. This way it doesn't
because a situation of fixing one after the other line, and having to
rerun the linter a bunch of times. Now all of the message body is
checked at once.

The message includes the line number as well so that it's easier to
figure out which line is the offending long line.