Add MessageEmptyFirstLine rule
This new rule checks if the message starts with an empty line. If not,
the first line of the message is considered part of subject.

The format it wants is this:

```
Subject line

First message line below an empty line.
```

The previous implementation with trim the message of whitespace and
empty lines before and after the message, but because this removes
important information about if the first line is empty or not, remove
this trimming.