Add musl build target along with artifact tests
Add the musl build that's compatible with Alpine Linux.

Add basic tests that test if the build artifacts actually run. It only
call `lintje --version`, which is enough to know if it compiled
correctly. In the future it should probably do more extensive testing,
like running the cargo tests, or some integration tests on those test
images.

The only target that's untested in this way is the
`aarch64-apple-darwin` target, which I can't run on any machine in an
automated way right now.