Ignore squash merge commits
When a commit is from a Pull Request merged with the squash strategy,
the commit ends with a number between parentheses. Ignore these commits
because they may violate some rules, but it's unwise to fix them after
they've been merged this way.

Improve logging about why a commit may be ignored.