Scan for emoji at the start of a subject
Add a SubjectPunctuation violation if the subject starts with an emoji.
There's no built-in Rust char method to check if a char or String is an
emoji, so I used the Regex crate's emoji detector. I think the "Emoji"
selector should be enough for now, there are other Emoji selectors as
well for `Emoji_Presentation`, `Emoji_Modifier`, `Emoji_Modifier_Base`
and `Emoji_Component`.