Check for more Unicode punctuation in subject
The `is_ascii_punctuation` check in Rust for characters doesn't include
all punctuation. Browsing some commits I see commits ending with
ellipsis and I want to check on that as well. I can't find a way to
check this automatically like `is_ascii_punctuation`, so keep track of a
list in the program itself for now of common Unicode punctuation to also
check on.