Add Windows build to the releases
Build a release of Lintje for Microsoft Windows using the GNU toolchain.
I can't use the native (MSVC) ABI used by Visual Studio as I'm not
compiling it on Windows itself, but cross compiling it on a Linux
container using cross.

Use the cross container to strip the build of debug symbols as well, to
reduce the final binary size. It was more effort to set up a new
container image to only strip the binary.

I've tested this on a Windows 10 machine using Powershell and Git for
Windows's Git Bash, both scenarios seem to work.