Upload debian packages to repository
Use Cloudsmith as a repository hosting for Debian installation packages.
This way users can install the Lintje package on Debian systems (like
Ubuntu) and receive updates this way too.

The Debian package building and uploading is done after the build and
as part of the release task. This way it cannot be accidentally
forgotten for new releases.

I've added Rake tasks that build and test the Debian packages as part of
the release process. In the next steps they will need to uploaded to a
hosting solution.