Improve rule messages for violations
Describe not what happened and what is wrong, but print what the user
should do to fix the violation.

For example the violation message "Subject·is·a·'Fix·test'·commit."
doesn't explain what the user should do, just what it detected. These
messages are some of the very first messages I didn't write to be more
helpful.