Update test script to work with new Rakefile build
Call the new build rake task to build the artifacts. The `script/build`
script has been replaced with the new rake build task.

Also update the `dist` directory for the archives. They have been moved
to a subdirectory in the `dist` directory

This broke because of an oversight from me. This script is not part of
the CI.