Add SubjectBuildTag rule to check for skip ci tags
The SubjectBuildTag checks if a commit subject contains a build tag.
These should be moved to the message body. The skip CI tag doesn't tell
anything about what kind of change was made. It's metadata only for the
CI system.

These tags are normally at the end of a subject like "Update README [ci
skip]", but the problem here is that people may think the punctuation
alone is the problem, not the tag.