Auto cancel builds on Semaphore on new push
Do not continue the build for branches that are not "main" that have
received a new push on them. I want to auto cancel those builds so that
it stops the old commit and starts building the new commit, speeding up
when I see the result of the latest push.