Add SubjectPrefix to recommend removing prefixes
Prefixes in commits don't say a lot about the change and the prefix
space should be used to explain the change in more detail.

This includes the first test that also checks the message of the
violation, because it includes the found match of the prefix. That was
added to the message to explain in more detail what should be removed.