Don't fail MergeCommit on local base branch merges
It's difficult to determine what type of merge the merge commit is
about. I want to warn about local branch merges into feature branches,
not about merges into the repository's main branch. Since Git has no
config for a "main" branch for a local repository this is difficult to
determine.

For this reason I'm narrowing the commits this rule triggers
on to only merges from remotes. I do want to trigger on local merges too
in the future, but those commits will be ignored for now. The tracking
issue for this is #6.

## Background

I notice two kind of merge commits being created by Git.

This subject is being created when I merge a local branch into the
repository's local "main" or "master" branches.

> Merge branch 'branch'

The following subject is created when a local branch is into another
local branch (that is not "main" or "master").

> Merge branch 'branch' into other

Perhaps this gives some hint for the future about how to distinguish
between the different types of merges, but I can't rely on it now, in
case a repository uses a different type of "main" branch name like
"trunk".