Fix GitLab merge commit detection
The GitLab UI doesn't show the full Merge Request reference
`org/repo!id`, instead it only shows the ID of the merge request.

The commit however does have the full reference, so I've updated the
commit detection to check for the full reference. Because this requires
a bit more complex detecting using variable names, I've used a Regular
expression to scan the body for this.

Fixes #8