Add command helper with better failure handling
Add the `run_command` function that checks the exit status of the run
command, not just if it failed to run at all.

This will hopefully fix #11, where running `git` on a system that does
not have `git` installed would not exit with an error message. It looked
like it was working, but it actually wasn't.

```
$ git
bash: git: command not found
$ lintje
0 commits inspected, 0 violations detected
```

This change also adds better error handling and error messages that will
highlight the #11 scenario and recommend users to check if the program
(most likely Git) was installed or not.

I only saw this problem on Docker emulation, when running AMD64 builds
on ARM64 or vice versa. Running the same architecture on the host and
Docker container would not result in this problem.