Expand cliche rule to cover plurals and more words
The original SubjectCliche rule only covered the "wip", "fix", "fix bug"
and "fix test" subjects.

It didn't cover plurals such as "fix tests". It does cover that with
this change.

Additionally I rewrote the simple string comparison to use a regular
expression here as well. This regular expression covers more words such
as "add", "update", "remove", etc. and different grammatical tenses such
as "adds", "adding", "added".

The updated implementation also doesn't check specifically only for
"test" or "bug" as the thing that can be fixed. If a subject matches
only two words with one of the prefix words it's considered a cliche.

All these examples don't say much about what change was really made:

- WIP
- Fix
- Fix bug
- Change config
- Update README
- Remove file
- Delete comments

## Alternative implementations

I considered splitting the subject string per word and matching the
first word and count if there were exactly two words in the subject.
This was a bit faster on subjects that didn't match, but was slower on
subjects that did match.

The regular expression has a more consistent execution time for common
subject lengths that was only a bit slower than the split string
implementation when it didn't match anything.

Fixes #12