Add --no-branch flag to disable branch validation
I've added the `--no-branch` flag for when the auto detected branch is
not accurate. I can imagine this being an issue in some CI environments
where the repository cloned is checked out by commit SHA and not the
branch.

There's no other way to disable branch names rules. Lintje could parse
the latest commits, or a selection of commits, for "lintje:disable"
comments, but that doesn't seem like the right place for it.

Part of #1