Add BranchNamePunctuation rule
The BranchNamePunctuation checks the branch for punctuation much like
the SubjectPunctuation rule.

It does not allow punctuation anywhere in the branch name. The
SubjectPunctuation only did not allow it at the start.

The branch name also can't start or end with Pu nutation much like the
subject.

Because these two rules have so much in common I extracted the
`is_punctuation` function to a `utils` module that can be reused by the
other modules.