Fix false positive emoji matches
The SubjectPunctuation rule checks on emoji, but it caught false
positives as well for the numbers 0-9, and symbols like * and #. This is
because of how Unicode works and how it's implemented in the Rust regex
library.

I've chosen to exclude the emoji using the ASCII codepoints to have less
false positives. I can revisit this later, but this seems like a good
solution for now.

For more information, see:
https://github.com/BurntSushi/ripgrep/discussions/1623#discussioncomment-28827