Add Debian installation method to CHANGELOG
This has been added in a previous commit, but was not included in the
CHANGELOG for the next release. It's actually already out for the
previous release of 0.3.1, but that was done much later after that
release. The next real release of the Debian installation packages will
be with the next update.

[ci skip]