Fix plural label printed on singular violation
If Lintje only finds one violation it would output "1 violations".
Instead print the singular version "1 violation" if this is the case.

I split up the printing of the line into two print statements to make
the arguments more manageable as more and more are added. It will still
only print one line.