Print number of ignored commits
When Lintje runs it may sometimes print it has validated 0 commits,
which can be confusing at times. It may appear like Lintje isn't
working, while instead it has ignored the commit, like a merge commit or
commit by a bot.

This change prints how many commits it has ignored if it has ignored any
to help take away any confusion here.

I've had to add an `ignored` field to the Commit struct, so it loops
over more commits now, but won't actually validate them. This is needed
so we can track how many commits were ignored to display it later.

If this becomes a (performance) problem it's possible to instead track
it in the `fetch_and_parse_commits` helper how many commits are ignored
and return it as part of a tuple.