Skip SubjectLength rule if it's a cliche
Sometimes there's a bit of noise in the violations that check different
things but one rule is more important and usually fixes the issue
already.

The SubjectCliche violation for a commit subject like "wip" should be
resolved first, and that usually means that the subject becomes more
detailed, longer, and thus won't be a problem for this SubjectLength
violation.

This way users see fewer violations and know which violation to focus
on.