Skip rules if it's a commit that needs rebasing
Sometimes there's a bit of noise in the violations that check different
things but one rule is more important and usually fixes the issue
already.

The NeedsRebase violation for a commit indicates that this commit
shouldn't be merged as is and is marked to be squashed into another
commit. The commit has prefix of "fixup!" or "squash!" set by Git, so we
shouldn't validate for a capital letter at the start of the subject.

These fixup or squash commits message body will be discarded when the
commit is squashed into another commit, so there's no need to add a
message body.

This way users see fewer violations and know which violation to focus
on.