Add DiffPresence rule to validate empty commits
The DiffPresence rule will check if a commit has any file changes
attached to it or not and add a violation when the commit is empty.

It is a basic check if there are any changes at all or not, it does not
check which changes have occurred.

## Supporting linting on existing commits

For the normal execution of lintje, I've updated the log command to
print the `shortstat` line after the message body per commit. The
`COMMIT_BODY_DELIMITER` was the best solution I could think off, the
split up the commit message body from the file changes of the commit.
I have no control over where Git prints the `shortlog`, so it had to be
printed after the commit message body.

## Supporting the commit-msg hook

For the commit message file option, I've had to run another command
using `git diff` to see if there any staged changes.

The commit message file was too unreliable as the user can customize
whether or not any file changes are listed there. The `--verbose` option
lists all changes, not just which files are listed, but `--quit` will
list nothing at all, which is impossible to parse.

So it runs another command to check the staged changes separately. Since
it's per execution of Lintje it should pose no performance impact. It
would be another matter if such a command would need to be run per
commit.