Remove error messages on empty commit subject
When a subject was empty some rules couldn't figure out the first
character of the subject, as there was none. This resulted in some
error messages being printed.

```
[ERROR] SubjectWhitespace validation failure: No first character found of subject.
[ERROR] SubjectCapitalization validation failure: No first character found of subject.
[ERROR] SubjectPunctuation validation failure: No first character found of subject.
[ERROR] SubjectPunctuation validation failure: No last character found of subject.
```

Ignore these rules, as they cannot validate anything when the subject is
empty. It makes sure the SubjectLength rule has been checked, in case
the rule was ignored in the commit.

Closes #17