Match more tags in the SubjectBuildTag rule
Instead of using a fixed list, match all "tags" that match the format of
"skip *" or "* skip". I don't know if the BUILD_TAGS list was complete
or not. Other systems may use other tags. As long as they match this
format, they should also match now.

This way other types of tags, like "[skip changeset]" or "[skip review]"
that we use at my employer, are also matched.

I left the "[no ci]" as is, matching it fully rather than matching the
"[no *]" format, as it's a little to generic and may match things I
didn't intend to.