Improve Unicode support for length based rules
Improve Unicode support for the SubjectLength and MessageLineLength
rules. Don't count the number of characters, but use the unicode-width
crate to determine the display width. The display width will be used as
the character length of the String, as that's what the user will see as
the String length.

```rust
// Using String.len()
println!("{:?}", "a̐".len());
// 3 (bytes)

// Using the previous method, counting characters
println!("{:?}", "a̐".chars());
// Chars(['a', '\u{310}'])
// This counts 2 characters. The accent is counted separately

// Using unicode-width
println!("{:?}", UnicodeWidthStr::width("a̐"));
// 1
```

The unicode-width crate implements the Unicode specification precisely,
but this also means it has some odd behavior (to me), see example below.
This means that certain emoji may be considered wider than other emoji,
and using one means the subject or message body line can contain
less/more characters than it actually should to trigger the rule.

```rust
println!("{:?}", UnicodeWidthStr::width("❤️"));
// 1
println!("{:?}", UnicodeWidthStr::width("✅"));
// 2
println!("{:?}", UnicodeWidthStr::width("👩‍🔬"));
// 4
```

I may add some more checks for this in the future, to return a more
accurate display width, by wrapping the unicode-width function.