Print better violation messages
The previous violation messages didn't have much design behind them.
They only showed the violation that occurred, but not where. If lucky,
it would print the offending character like in the SubjectPunctuation
rule.

```
$ lintje
Fix bug [skip ci]
  SubjectPunctuation: Remove punctuation from the end of the commit subject: ]
```

## New messages

The new violation messages that are printed provide more context to the
violation that occurred. It will show the location where the problem was
found and will provide hints about how to resolve the issue.

```
$ lintje
SubjectPunctuation: The subject ends with a punctuation character: `]`
  d3349a7:1:17: Fix bug [skip ci]
    |
  1 | Fix bug [skip ci]
    |                 ^ Remove punctuation from the end of the subject
```

This message shows first which rule was broken, with an error message.
Then it prints the commit SHA, the line and column on which the problem
was found. It prints the commit subject after that so users can easily
find back which commit it was in.

The hint section below will highlight the character, or range of
characters, the rule triggered on. The hint message will tell the user
how to resolve it.