Fix tab rendering by rendering it as spaces
Tabs render different widths based on position, this makes it difficult
to underline it accurately and not accidentally highlight unrelated
characters.

To fix the render width, render the tab as four spaces. That's the same
width Lintje already uses to calculate the tab width. Since I only
replace the tabs in the final rendered output, it should not affect the
rules that calculate presence and position of the tab character or other
character after it.

I think other apps do the same, because if I add a tab to a subject to
trigger the SubjectWhitespace rule tools like `git log` will render the
tabs as spaces as well. Only `git log` render it as eight spaces.