Improve rendering of certain Unicode characters
Not all characters have the same display width. The unicode-width crate
returned the display width of some emoji, especially those with
modifiers on them, as having a larger display width than they would
actually render as. This is because it appears to includes every
modifier character's width as well, even though they were rendered as
one combined entity.

I added some exceptions in the display_width function to determine the
width that is rendered in the terminal, which is mostly two columns.

This way the new violation messages have their hints accurately aligned
under the characters that were the problem.

This function may need some tweaks in the future as more emojis and
characters are used and it is reported that the hints are not properly
aligned.