Refactor output formatters
Move all the output formatters to their own formatter module. This way
it's not spread over the main file and Violation struct.

Initially it was easy to add the `formatted_context` method on the
Violation struct, but I want to add things like color to the output, and
I don't want the Violation struct to have to know about that setting.