Fix string width calculation in rules & formatter
When a line contained an emoji before a marked area in a line the
formatter would think that leading substring was two characters long.
This was because of some incorrect usage of the `display_width`
function.

The formatter and some rules used the `display_width` function
incorrectly. It's meant to check character length and not strings like
whole lines. Make the behavior of the `display_width` function private
as `display_width_char` and implement `display_width` as previously
used, to calculate a string length.

This implementation is similar to `line_length_stats`, but I don't use
the majority of the return values there, so I implemented it again.