Adjust Unicode width calculation
Remove the "fix" for Unicode characters that use the variation selector
16 character. This is about characters like the heart emoji, the cloud,
the umbrella, etc. see the examples in the tests.

Lintje considered all these characters to have a width of two columns.
This is kind of correct, but not how most terminals will display those
emoji. So while Lintje calculated it correctly, the formatter's output
would be misaligned because the app used to render it calculates it as
only one column.

I removed the correction so the hints in violations are properly aligned
again. If I want to be very correct I could also inject an space after
these emoji to have them render as two columns, but that might break
other things again.