Skip checks if MergeCommit or NeedsRebase is found
Ignore other rules if a commit has a MergeCommit or NeedsRebase
violation. When these violations occur the commit needs to be rebased,
so any other issues will hopefully be resolved in the rebase, such
missing message body, or subject length.

This will reduce the number of violations printed and helps the user
focus on the important violations.

If the rules are skipped, these violations will not be added and the
other rules will trigger a validation.