Ignore the scissor line in Git commit hook file
The Git scissor line (for cleanup mode scissor) will not be interpreted
as the commit subject line, if it's the first line in a Git commit hook
file.

It will instead consider the commit as having an empty subject and
message body. This will prevent any unexpected violations on the
scissor line when the Git commit process is aborted by removing the
subject and message body from the Git commit message file.