Improve support for Git commit verbose mode
When `git commit` is called with `--verbose` it adds a scissor line to
the default commit file content even if the `commit.cleanup` mode isn't
`scissors`.

Git does not include the scissor line and anything after the scissor
line, in any mode, in the commit message.

Even in verbatim mode, that should include everything in the commit
message file, the content after the scissor line is ignored. However the
default comments above the scissor line _are_ included.

It looks like the scissor line doesn't only apply to scissors mode, but
for all modes. I've changed the Git commit message parser to now scan
for the scissor line in every mode and stop parsing the subject and
message body when it encounters this line.

Fixes #21