Improve leading empty line detection
Don't consider leading empty lines as part of the subject for every
cleanup mode except Verbatim.

Extract the line cleanup logic to a separate function so it can also
apply to subject lines. This way it also trims of any trailing
whitespace characters so that they don't count towards the line length.

It now also ignores comments on the first line in the "strip" cleanup
mode, because of this shared function.

Update tests to test trimming trailing whitespace all the time, so I
don't need to write separate tests for it.