Add MessageTicketNumber hint
The MessageTicketNumber hint will provide a hint to users that they do
not have a ticket number mentioned in the message body.

This hint is the first hint added to Lintje. Whenever Lintje only prints
hints like this one and others in the future, it will not fail the
linting process. These hints do not need to be resolved.

Hints can be disabled entirely with the `--no-hints` flag.