Add addition type context to issue context
I don't only want to show hints separately as their own issue, but also
for error issues. This provides more useful context with the actual
contents of the problematic content.

It will now print something like this:

```
SubjectTicketNumber: The subject contains a ticket number
  4225bcc:1:10: Fix bug. Closes #123
    |
  1 | Fix bug. Closes #123
    |          ^^^^^^^^^^^ Remove the ticket number from the subject
  3 | Closes #123
    | ----------- Move the ticket number to the message body
```

Where `Closes #123` is the actual content present in the commit.

Previously it printed this:

```
SubjectTicketNumber: The subject contains a ticket number
  4225bcc:1:10: Fix bug. Closes #123
    |
  1 | Fix bug. Closes #123
    |          ^^^^^^^^^^^ Move the ticket number to the message body
```

There still needs to be some kind of separator between the line on which
an issue was detected and the line for which the addition is present.