Add context line separator between different lines
When there's context that's for different lines that do not connect,
like line number 1 and line number 10, add a small separator to indicate
a break in context.

This will help in indicating some context is missing between these two
context lines. I don't want to show everything between these two
context, because it can be many lines of commit message.

This is currently only visible in the SubjectTicketNumber rule output as
the `~~~` separator:

```
SubjectTicketNumber: The subject contains a ticket number
  4225bcc:1:10: Fix bug. Closes #123
    |
  1 | Fix bug. Closes #123
    |          ^^^^^^^^^^^ Remove the ticket number from the subject
   ~~~
  3 | Closes #123
    | ----------- Move the ticket number to the message body
```