Change sort order of flags in --help output
Print the flags in the order they are defined. This way the flags that
are about the same kind of setting, like `color` and `no-color` are
listed below one another.

I've also changed the declaration of some options to print the options
users will probably need most at the top.

Use one of clap's settings, DeriveDisplayOrder, to accomplish this.