Add support for options file
Allow users to configure a Lintje options file. This file will be used
as the base options and flags given to the Lintje executable.

The options file has one or multiple flags and options per line. It also
supports commented lines starting with the number symbol (`#`).

```
--color
# Disable hints
--no-hints

# Disable branch validation
--no-branch
```

The location of this file can be configured with the
`LINTJE_OPTIONS_PATH` environment variable.

Closes #32