Check presence of "Part of #<ticket>" in message
For the MessageTicketNumber rule, check for "Part of #<ticket number>"
sub string. I use this myself a bunch to link back to issues the commit
shouldn't necessarily close when merged into the default branch.

I used the RegexBuilder for this matcher because I want it to match case
insensitive, something the other regular expressions should really do as
well.

Closes #40