Add test helper for asserting formatter output
It was getting annoying to keep having to match the output exactly from
the formatter. This helper doesn't care about how much a line is
indented or not, that's up to the formatter's test suite to ensure.

It doesn't require the first line, which always empty, to be asserted.
If any lines are after the asserted lines, this is ignored.