Print link to help under every issue
When an issue is found Lintje tries to print as much help as it can,
along with some suggestions. Sometimes this isn't enough. I've added a
link to the docs with more explanation and examples may help the user
further.

It can also be printed at the very end or in some non-duplicate way, but
this can be revised later.

I also thought about using a link shortening service like
`r.lintje.dev/RuleName` to make sure the links will always work, even if
the docs structure changes, but this also can be added later if
necessary.

Closes #35