Add a way to add gaps manually
When an issue context is printed and I want to show multiple things,
like part of the diff and the message, allow for a way to indicate the
split manually.

The gap context type will add a split using the scissors. These scissors
are also used to indicate large gaps between lines in the message, like
line 1 and line 10.