Rename NeedsRebase rule to RebaseCommit
The RebaseCommit name matches the naming with MergeCommit rule better.
This helps to indicate the type of commit rather than the action that
needs to be taken to resolve it.

The NeedsRebase name can still be used as an alias to disable the
RebaseCommit rule with `lintje:disable NeedsRebase`. This alias will be
removed eventually.