Move error handling to one location
Instead of checking for commit and branch errors separately, move this
to the `handle_command` function, which replaces the already hardly used
`handle_result` function. It was unused because of that separate error
handling.

The downside is that you can't see branch validation results when commit
validation failed, but that probably never happens and if it does you
would want to know about that more so than the branch validation.