Make error messages more human friendly
Instead of showing all details from an error by default, put it behind
the `--debug` flag.

By default Lintje will now only print the error message from the
command that failed. If it ran `git log` in a directory that's not a
Git repo, it will now print the message below instead of the detailed
message.

> Lintje encountered an error: fatal: not a git repository (or any of
> the parent directories): .git

Closes #45