Refactor rule to use RuleValidator trait
Use a trait to have a shared function signature to call `validate` on
each Rule implementation struct using the RuleValidator trait.

This way I don't have to call validate in every match-arm of the enum
matcher in Rule. I do have to wrap it in a Box because the size of each
struct isn't known at compile time.

This does allow me to add more generic behavior on top of rule structs
without having to do this in every match-arm.