Add a rule dependency check
Make sure that rule A, that rule B depends on, is validated before
rule B is validated. Otherwise it can never catch those cases.

This is a runtime check and the overhead is minimal. Ideally this is
caught at compile time or in a test case, but I don't want to run every
rule validation for every test like it was done in the past.