Refactor formatter to allow error without message
In the MessagePresence I want to highlight some rules with underlines
without printing a message behind that underline like Lintje does
everywhere else. That hint message is only shown on the last
line on which the error occurred, and it doesn't need to repeat it for
every error line.

Previously I solved this with passing an empty string as the message,
but since it's an Option already, set None as the message instead.