Don't count ticket references to message length
When a commit message is only ticket references don't count it as valid.

The ticket references are not counted towards the message body length,
even if they are in the middle of a line, not just if they are the only
thing on a line.

Lintje still wants the commit to have some substantial message body.