Add hint to add changeset file to commit
Changesets are useful files to generate a changelog whenever a new
release is made without having to deal with changelog file merge
conflicts.

Add a rule to hint about adding a changeset file when it's not part of
the commit.

The file path it suggests is not the standard path used in
https://github.com/changesets/changesets, but what I use at work. I'll
try to base this on one of the actual detected paths later on so it's
more context aware.

The changelog entry already mentions it will only check this rule when
a changeset directory is present, but it does not do that yet.

Closes #38