Move Git commit validation to main
Move it out of the `commit_for` function that creates the commit object.
I also don't want to validate the commit on creation of the struct
object, so I'm not adding it to `Commit::new`.

In the future I want to be able to pass some arguments to the Commit
`validate` method and I don't want to have to pass that through every
function until it reaches `commit_for`. Instead call `validate` in the
loop that shows the results.

I should do the same refactor for branch, but that can be done later.