Only validate changesets when present in Git repo
The MessageChangesetTag rule only needs to be triggered when a
repository uses changesets. For repositories that don't this rule will
be confusing.

I've added a ValidationContext struct which can contain more of this
detail about the repo context in the future. I want to automate setting
these details and which rules are validated, and not introduce some kind
of config file to enable and disable rules.