Support trailer parsing for git log
When Lintje fetches and parses commits from `git log`, ask Git what the
trailers are so they are easier to remove from the message body.

This can't be done for the commit hook formatter, because no commit
exists yet and I can't ask Git. I'll need to implement that manually.